Moved ExecutionBlocker to graph_utils.py, maintained backwards compatibility
